+++ src/components/day-list.js
@@ -1,66 +1,12 @@
 import React, { useState, useEffect } from 'react'
 
-import { useQueryParam, StringParam, BooleanParam } from 'use-query-params';
-
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faCaretSquareRight } from '@fortawesome/free-regular-svg-icons'
+import { stateFromQueryParam, StringParam, BooleanParam }
+  from '../components/utils'
 
 import appendQuery from 'append-query'
 
-import Avatar from "./avatar"
-import TimeLabel from './time-label';
+import EventCard from "./event-card"
 import ControlBox from './control-box';
-
-const EventCard = ({
-      name, description, isGovTeam, eventType,
-      duration, videoUrl, isChairman
-    }) => {
-    const color = isChairman ? "white": "black"
-    const background = isChairman ? "black": "white"
-    return <div style={{
-      width: "100%",
-      border: "2px solid black",
-      borderTopRightRadius: "5px",
-      borderBottomRightRadius: "5px",
-      color: color,
-      background: background,
-    }}>
-      <div style={{padding: "5px", position: "relative"}}>
-          <div style={{
-            float: "left",
-            width: isChairman ? `0`: `10px`,
-            height: "100%",
-            background: isGovTeam ? `#0E64B9` : `#E1161F`,
-            position: "absolute",
-            top: 0,
-            left: 0
-           }}></div>
-          <div style={{float: "right", fontWeight: "bold"}}>
-              {eventType} <TimeLabel duration={duration}/>
-          </div>
-          <div style={{float: "left", marginLeft: "15px"}}>
-            <div style={{float: "left", marginRight: "10px"}}>
-              <Avatar src={name} width={isChairman ? 40 : undefined}/>
-            </div>
-            <div style={{float: "left", fontSize: "1.2rem"}}>
-                <b>{name}
-                  <a href={videoUrl} target="_blank"
-                    style={{
-                        marginLeft: "5px",
-                        color: color,
-                        textDecoration: "none",
-                    }}
-                  >
-                    <FontAwesomeIcon icon={faCaretSquareRight}/>
-                  </a>
-                </b><br/>
-                {!isChairman && description}
-            </div>
-          </div>
-          <div style={{clear: "both"}}></div>
-      </div>
-    </div>
-}
 
 const short_months = ["", "ม.ค.", "ก.พ.", "มี.ค.", "เม.ย.", "พ.ค.", "มิ.ย.", "ก.ค.", "ส.ค.", "ก.ย.", "ต.ค.", "พ.ย.", "ธ.ค."];
 const thai_date = (str) => {
@@ -68,21 +14,8 @@ const thai_date = (str) => {
   return `${+dmy[0]} ${short_months[+dmy[1]]} ${+dmy[2] + 543}`;
 }
 
-const stateFromQueryParam = (name, type) => {
-    const [valueFromUrl, setQueryToUrl] = useQueryParam(name, type)
-    const [stateValue, setStateValue] = useState(valueFromUrl)
-
-    useEffect(() => {
-      setQueryToUrl(stateValue)
-    }, [stateValue])
-
-    return [stateValue, setStateValue]
-}
-
 const DayList = ({data}) => {
-
     const [query, setQuery] = stateFromQueryParam('q', StringParam)
-
     const [chairmanFilter, setChairmanFilter] = stateFromQueryParam(
       'chariman',
       BooleanParam
